Used jQuery to Display Loading Notices in Nav until Loaded.
Added ID's to Links in the Nav bar.
Added relevant Notice for each link. Hide them by default.
Added jQuery On-Click Events to each Link ID in the DOM.
When user clicks a one of the links in the Nav Bar (Home, Student, Job, Blog, or Project), the following occurs:
 - jQuery On-Click event for that DOM ID/Class name is triggered.
 - 'Loading...' notice associated with Link being loaded is displayed immediately, and all other notices hidden to prevent overlap.
 - Once the page associated with the link clicked is loaded, the '$(document).ready' function is triggered for that page, causing the 'Loading...' notice to disappear again.

Does not display notice if user visits the URL directly without clicking the Nav Bar links (not expected to occur anyway).