Refactored jQuery Notices. Resolved issue of multiple notices appearing.
Resolved issue when testing where all Notices contents would appear sometimes (jQuery hide not being triggered for some reason).
Now instead of the Notices relying on the .hide() to hide them and their contents by default on page load, the paragraph tags
of these Notices are now empty by default. On-Click events on the Nav bar links trigger the respective Notice tag to appear and the event passes the associated contents of the notice message to the innerHTML of the DOM element.