Fixed Skills Animations with jQuery to be Mobile First compatible.
Problem was that Skills Animations contents were not loading correctly for single column grid (screen sizes less than 515px).

Tested numerous methods with jQuery to adjust the margin of the contents when the user adjusted the browser screen size (without reloading the webpage). Unfortunately none of these solutions worked (animation margin doubled each time window was resized, unable to understand why).

Solution was achieved with using last resort of triggering a reload of the homepage when the user choses to adjust the browser screen size. This works and avoids any problems with animation conflicts.

Useful links used to get it to work were as follows:
http://stackoverflow.com/questions/4298612/jquery-how-to-call-resize-event-only-once-its-finished-resizing